Remove inappropriate IDBKeyRange usages from lib.dom.d.ts.
I hit a compile error because IDBCursor.primaryKey was incorrectly typed as
`IDBValidKey | IDBKeyRange` instead of just `IDBValidKey`. I fixed it and
manually audited the rest of the instances of `IDBValidKey | IDBKeyRange` and
fixed them.  If there's a better way to fix these other than manual audit based
on my understanding of the IndexedDB API, let me know.
-