Change SubtleCrypto method return types to Promise (#843)
Previously, these methods returned PromiseLike, even though the spec has
always required them to return real Promises and browsers have always
done so. According to #60, this was done in order to work around #47;
now that #47 has been fixed, this workaround is obsolete.

This generally should not be a breaking change since it is making the
SubtleCrypto type contract stricter and that API is generally
implemented by the platform, not by users.